Fix Path of Force
