+++ README.md
@@ -24,7 +24,7 @@ This project aims to supply, over time, examples and best practices of PLCnext T
 |[04](Examples/ThreadExample/)| [Thread Example](Examples/ThreadExample/README.md)| This describes how to implement Threads in a component of C++ project.| [SourceCode](Examples/ThreadExample/src/)
 |[05](Examples/DataAccess/)| [Data Access Example](Examples/DataAccess/README.MD)| This example shows how to read and write GDS variables using the Data Access service.| [SourceCode](Examples/DataAccess/src/)
 |[06](Examples/SubscriptionExample/)| [Subscription Example](Examples/SubscriptionExample/README.MD)| This example shows how to read GDS variables using the Subscription service.| [SourceCode](Examples/SubscriptionExample/src/)
-|[07](Examples/ForceExample/)| [Force Example](Examples/ForceExample/README.MD)| This example shows how GDS variables can be forced using the Force service.| [SourceCode](Examples/ForceExample/src/)
+|[07](Examples/Force/)| [Force Example](Examples/Force/README.MD)| This example shows how GDS variables can be forced using the Force service.| [SourceCode](Examples/Force/src/)
 |[08](Examples/NotificationExample/)| [Notification Example](Examples/NotificationExample/README.MD)| This example shows how to use the Notification Manager.| [SourceCode](Examples/NotificationExample/src/)
 |[09](Examples/IncludeOpenSourceLibrary/)| [Include an Open Source Library](Examples/IncludeOpenSourceLibrary/README.md)| This describes how to use a third-party open-source library in a C++ project that is built for PLCnext Control using Eclipse|
 |[10](Examples/FileStreamExample/)| [Commons::Io::FileStream Example](Examples/FileStreamExample/README.md)| This Example shows how to use a File Stream to check the binary generation timestamp whenever a Component is reinitialized.|[SourceCode](Examples/FileStreamExample/src/)
@@ -66,8 +66,8 @@ There are different ways to build your code:
 1. Modify the [build.sh](tools/build.sh) script to fit your needs
     1. select hardware
     1. select project
-    1. rename folder to fit namespace name (e.g. mv Examples/DataAccessExample Examples/DataAccess)
-    1. switch between acfproject and pwoject
+    1. rename folder to fit namespace name (e.g. mv Examples/OpcUaMethodExample Examples/OpcPlcManager)
+    1. switch between `acfproject` and `project` command if necessary
 1. Make sure it is executable for example with ```chmod +x tools/build.sh```
 1. Execute the script ```./tools/build.sh``` or  ```sh tools/build.sh ```
